Putting everything to the norm
hopefully, i broke nothing.
