+++ srcs/exec.c
@@ -104,7 +104,8 @@ int	execute_single_command(t_simple_command *commands)
 		if (!handle_redirections(commands->redirections))
 		{
 			if (is_builtin(arguments[0]))
-				set_status_code(execute_builtin(arguments[0], &arguments[1]), true);
+				set_status_code(execute_builtin(arguments[0], &arguments[1]),
+					true);
 			else if (find_exe_path(arguments[0]))
 				execute_command(arguments);
 			else
